requirements.txt
fastapi
uvicorn
torch
torchvision
pandas
Pillow
ultralytics
seaborn
python-multipart
jupyter_client
jupyter_core